add redis storage

+++ config.go
@@ -19,5 +19,5 @@ type WebApiCfg struct {
 type RedisCfg struct {
 	Host string `toml:"host"`
 	Pass string `toml:"pass"`
-	Db   int    `toml:"db"`
+	Db   int64  `toml:"db"`
 }
